OneSignalSDK.js is now a shim to load ES6 vs ES5
* This is done to optimize bandwidth used by loading the SDK.
* OneSignalSDK.js will now load the full ES6 SDK if push is supported.
   - If push is not supported a ES5 version will be loaded with stubbed SDK methods.
   - iOS for example will be loading this small stub SDK
* If OneSignalSDK.js is imported into a service worker it will load OneSignalSDKWorker.js
* Moved internal isPushNotificationsSupported implementation into it's own file BrowserSupportPush.ts
  - This was needed to keep the shim from adding a bunch of unused functions.
* New size results before gzip
  - OneSignalSDK.js        - 9.95 KiB
  - OneSignalPageSDKES6.js - 208 KiB
  - OneSignalPageSDKES5.js - 1.94 KiB
  - OneSignalSDKWorker.js  - 108 KiB
* Possible ~8 KiB size saving if we remove isPushNotificationsSupported() + browser lib from PageSDKES6
