+++ package.json
@@ -20,12 +20,13 @@
     "bundle": "TESTS=false $(yarn bin)/webpack --config build/config/webpack.config.js",
     "bundle-sw": "TESTS=false $(yarn bin)/webpack --config build/config/serviceworker.config.js",
     "bundle-sdk": "TESTS=false $(yarn bin)/webpack --config build/config/sdk.config.js",
+    "bundle-page-sdk-es6": "TESTS=false $(yarn bin)/webpack --config build/config/page-sdk-es6.config.js",
     "bundle:analyze": "ANALYZE=true ENV=production TESTS=false $(yarn bin)/webpack --config build/config/webpack.config.js --profile --json > stats.json",
     "bundle-sw:analyze": "ANALYZE=true ENV=production TESTS=false $(yarn bin)/webpack --config build/config/serviceworker.config.js --profile --json > stats.json",
     "bundle-sdk:analyze": "ANALYZE=true ENV=production TESTS=false $(yarn bin)/webpack --config build/config/sdk.config.js --profile --json > stats.json",
-    "build:dev": "ENV=development yarn transpile:sources && ENV=development yarn bundle-sw && ENV=development yarn bundle-sdk && ENV=development build/scripts/publish.sh",
-    "build:staging": "ENV=staging yarn transpile:sources && ENV=staging yarn bundle-sw && ENV=staging yarn bundle-sdk && ENV=staging build/scripts/publish.sh",
-    "build:prod": "ENV=production yarn transpile:sources && ENV=production yarn bundle-sw && ENV=production yarn bundle-sdk && ENV=production build/scripts/publish.sh",
+    "build:dev": "ENV=development yarn transpile:sources && ENV=development yarn bundle-sw && ENV=development yarn bundle-sdk && ENV=development yarn bundle-page-sdk-es6 && ENV=development build/scripts/publish.sh",
+    "build:staging": "ENV=staging yarn transpile:sources && ENV=staging yarn bundle-sw && ENV=staging yarn bundle-sdk && ENV=staging yarn bundle-page-sdk-es6 && ENV=staging build/scripts/publish.sh",
+    "build:prod": "ENV=production yarn transpile:sources && ENV=production yarn bundle-sw && ENV=production yarn bundle-sdk && ENV=production yarn bundle-page-sdk-es6 && ENV=production build/scripts/publish.sh",
     "test": "ENV=development yarn transpile:tests && ava --verbose --color --watch",
     "test:noWatch": "ENV=development yarn transpile:tests && ava --verbose --color --watch=false",
     "publish": "yarn clean && yarn build:prod && yarn "
